Refactor package structure and update imports for core components
