--- exekutor-core/src/main/kotlin/com/yonatankarp/exekutor/core/StepResult.kt
@@ -1,12 +1,0 @@
-package com.yonatankarp.exekutor.core
-
-data class StepResult(
-    val outcome: Outcome,
-    val details: Any? = null,
-)
-
-enum class Outcome {
-    PASS,
-    FAIL,
-    FRICTION_REQUIRED,
-}
