+++ exekutor-core/src/main/kotlin/com/yonatankarp/exekutor/core/api/ExecutionContext.kt
@@ -1,4 +1,4 @@
-package com.yonatankarp.exekutor.core
+package com.yonatankarp.exekutor.core.api
 
 interface ExecutionContext {
     val timeBudgetMs: Long
