+++ exekutor-core/src/main/kotlin/com/yonatankarp/exekutor/core/api/Outcome.kt
@@ -1,0 +1,7 @@
+package com.yonatankarp.exekutor.core.api
+
+enum class Outcome {
+    PASS,
+    FAIL,
+    FRICTION_REQUIRED,
+}
