+++ exekutor-core/src/main/kotlin/com/yonatankarp/exekutor/core/api/Step.kt
@@ -1,0 +1,7 @@
+package com.yonatankarp.exekutor.core.api
+
+interface Step<C : ExecutionContext> {
+    val name: String
+
+    suspend fun execute(context: C): StepResult
+}
