+++ exekutor-core/src/main/kotlin/com/yonatankarp/exekutor/core/api/StepResult.kt
@@ -1,0 +1,6 @@
+package com.yonatankarp.exekutor.core.api
+
+data class StepResult(
+    val outcome: Outcome,
+    val details: Any? = null,
+)
