--- exekutor-core/src/main/kotlin/com/yonatankarp/exekutor/core/steps/DefaultPlanBuilder.kt
@@ -1,8 +1,0 @@
-package com.yonatankarp.exekutor.core.steps
-
-import com.yonatankarp.exekutor.core.ExecutionContext
-
-fun <C : ExecutionContext> defaultPlanBuilder(): (C) -> List<Step<C>> =
-    {
-        StepRegistry.all().filterIsInstance<Step<C>>()
-    }
