--- exekutor-core/src/main/kotlin/com/yonatankarp/exekutor/core/steps/StepRegistryDsl.kt
@@ -1,9 +1,0 @@
-package com.yonatankarp.exekutor.core.steps
-
-import com.yonatankarp.exekutor.core.ExecutionContext
-
-class StepRegistryDsl {
-    fun <C : ExecutionContext> step(factory: () -> Step<C>) {
-        StepRegistry.register(factory())
-    }
-}
