+++ exekutor-sample/src/test/kotlin/com/yonatankarp/exekutor/sample/ExekutorSampleTest.kt
@@ -1,7 +1,7 @@
 package com.yonatankarp.exekutor.sample
 
-import com.yonatankarp.exekutor.core.ExecutionDecision
-import com.yonatankarp.exekutor.core.StepExecutionEngine
+import com.yonatankarp.exekutor.core.api.ExecutionDecision
+import com.yonatankarp.exekutor.core.engine.StepExecutionEngine
 import kotlinx.coroutines.test.runTest
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.Test
